requirements.txt
torch
torchvision
tqdm
open_clip_torch
ftfy
numpy
pandas
Pillow
regex
scipy
six